+++ SwitchDash 3.4/Engine/src/main/java/es/ucm/vdm/engine/State.java
@@ -8,8 +8,8 @@ public abstract class State {
         this.game_ = game;
     }
 
-    public abstract void update(float deltaTime);
-    public abstract void render(float deltaTime);
+    public abstract void update(double deltaTime);
+    public abstract void render(double deltaTime);
     public abstract void pause();
     public abstract void resume();
     public abstract void dispose();
